requirement.txt
mysql-connector-python==8.0.33
mysqlclient==2.2.0
protobuf==3.20.3
xlrd==2.0.1